Update attendance entity and files (user / meeting entity)
