+++ src/enums/attendance-status.enum.ts
@@ -1,7 +1,8 @@
+
 export enum AttendanceStatus {
-  PRESENT = 'present',
-  ABSENT = 'absent',
-  LATE = 'late',
-  EXCUSED = 'excused',
-  PENDING = 'pending',
+  PENDING = 'pending',    // Default state (student self-marked)
+  PRESENT = 'present',    // Teacher-confirmed
+  ABSENT = 'absent',      // Teacher-confirmed
+  LATE = 'late',          // Teacher-confirmed
+  EXCUSED = 'excused',    // Teacher-confirmed
 }
