+++ src/users/entities/user.entity.ts
@@ -1,8 +1,9 @@
-import { Column, Entity, PrimaryGeneratedColumn, OneToMany, OneToOne } from 'typeorm';
+import { Column, Entity, PrimaryGeneratedColumn, OneToMany, OneToOne, ManyToOne } from 'typeorm';
 import { Role } from './role.entity';
 import { Teacher } from 'src/teacher/entities/teacher.entity';
 import { Comment } from 'src/comments/entities/comment.entity';
-;
+import { Classes } from 'src/classes/entities/classes.entity';
+import { Attendance } from 'src/attendance/entities/attendance.entity';
 
 @Entity()
 export class User {
@@ -32,4 +33,13 @@ export class User {
 
     @OneToMany(() => Comment, comment => comment.user)
     comments?: Comment[];
+
+    @ManyToOne(() => Role, role => role.user)
+    role: Role;
+
+    @ManyToOne(() => Classes, classes => classes.teacher, { nullable: true })
+    class?: Classes;  // Only for students/teachers
+
+    @OneToMany(() => Attendance, attendance => attendance.user)
+    attendances: Attendance[];
 }
